+++ guillotina/fields/__init__.py
@@ -1,3 +1,4 @@
 from guillotina.fields.annotation import BucketListField  # noqa
 from guillotina.fields.dynamic import DynamicField  # noqa
+from guillotina.fields.files import CloudFileField  # noqa
 from guillotina.fields.patch import PatchField  # noqa