+++ guillotina/files/__init__.py
@@ -3,7 +3,6 @@ from .const import CHUNK_SIZE  # noqa
 from .const import MAX_REQUEST_CACHE_SIZE  # noqa
 from .const import MAX_RETRIES  # noqa
 from .field import BaseCloudFile  # noqa
-from .field import CloudFileField  # noqa
 from .manager import FileManager  # noqa
 from .utils import convert_base64_to_binary  # noqa
 from .utils import get_contenttype  # noqa